requirements.txt
langchain_core==0.3.50
langchain_ollama==0.3.0
matplotlib==3.10.1
neo4j==5.24.0
networkx==3.2.1
numpy==2.2.4
pandas==2.2.3
PuLP==2.9.0
py_stopwatch==0.2.2
requests_cache==1.2.1
scipy==1.15.2
seaborn==0.13.2
tqdm==4.66.4
tqdm==4.66.5
